+++ lib/shelley/src/Cardano/Wallet/Shelley/Network.hs
@@ -27,7 +27,7 @@ module Cardano.Wallet.Shelley.Network
     , withNetworkLayer
 
       -- * Logging
-    , NetworkLayerLog
+    , NetworkLayerLog (..)
     ) where
 
 import Prelude
